requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
PyMySQL==1.1.2
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
